+++ lib/basic/widgets.dart
@@ -24,3 +24,4 @@ export 'strait_widget.dart';
 export 'parsable.dart';
 export 'indexed_file.dart';
 export 'raw_file.dart';
+export 'predicate.dart';